backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
PyJWT==2.9.0
pandas==2.2.2
openpyxl==3.1.5
watchdog==4.0.1
httpx==0.27.0
sse-starlette==2.1.0
slowapi==0.1.9
aiofiles==24.1.0
